# Reduce serialization during Downstairs reads