--- a/protocol/src/lib.rs
+++ b/protocol/src/lib.rs
@@ -677,11 +677,6 @@ impl Message {
 /// Message to be sent down the wire
 #[derive(Debug)]
 pub enum WireMessage<M> {
-    /// Normal message to be sent down the wire
-    ///
-    /// This is serialized with the existing [`CrucibleEncoder`]
-    Message(Message),
-
     /// Pre-serialized message to be sent down the wire
     ///
     /// This is sent by sending
@@ -691,6 +686,9 @@ pub enum WireMessage<M> {
     ///
     /// The values of `M` and the byte array must match the equivalent
     /// [`Message`] serialized with a [`CrucibleEncoder`].
+    Message(Message),
+
+    /// Pre-serialized message to be sent down the wire
     RawMessage(M, bytes::Bytes),
 }