(This PR is staged on top of #1182, and is therefore marked as a draft; you should ignore all changes in `crucible-protocol` and `crucible-upstairs`)

#1087 changed how we handle writes Upstairs to avoid memcpy'ing bulk data in `Message::Write/WriteUnwritten`.  This PR does the same thing for `Message::ReadResponse` in the Downstairs.

Just like #1087, it's Unfortunately Messy: we have to do the `bincode` encoding by hand at the same time that we're reading data from the file, to avoid an extra copy operation.

The benefits are what you'd expect: the 11% of our runtime spent in `memcpy` disappears, and things get faster.

Here's the flamegraph before these changes...
![ds_main](https://github.com/oxidecomputer/crucible/assets/745333/8c012922-1180-4009-9d44-d71bd6ca525d)
([interactive version](https://github.com/oxidecomputer/crucible/assets/745333/8c012922-1180-4009-9d44-d71bd6ca525d))

...and after
![ds_zc](https://github.com/oxidecomputer/crucible/assets/745333/45fb7415-9789-4a76-9751-7020e6fe3dc7)
([interactive version](https://github.com/oxidecomputer/crucible/assets/745333/45fb7415-9789-4a76-9751-7020e6fe3dc7))

Running my usual `fio` sweep, I see noticeable improvements in large read performance.

### Before
```
4K READ: bw=27.8MiB/s (29.1MB/s), 27.8MiB/s-27.8MiB/s (29.1MB/s-29.1MB/s), io=1667MiB (1748MB), run=60003-60003msec
1M READ: bw=525MiB/s (550MB/s), 525MiB/s-525MiB/s (550MB/s-550MB/s), io=30.8GiB (33.0GB), run=60048-60048msec
4M READ: bw=596MiB/s (625MB/s), 596MiB/s-596MiB/s (625MB/s-625MB/s), io=35.0GiB (37.6GB), run=60138-60138msec
```
### After
```
4K READ: bw=27.8MiB/s (29.2MB/s), 27.8MiB/s-27.8MiB/s (29.2MB/s-29.2MB/s), io=1670MiB (1751MB), run=60027-60027msec
1M READ: bw=675MiB/s (708MB/s), 675MiB/s-675MiB/s (708MB/s-708MB/s), io=39.6GiB (42.5GB), run=60035-60035msec
4M READ: bw=756MiB/s (793MB/s), 756MiB/s-756MiB/s (793MB/s-793MB/s), io=44.4GiB (47.7GB), run=60127-60127msec
```
(The Upstairs includes the changes from #1181, which is why initial performance is slower than you may expect)