@mkeeter 2024-03-13
#1206 does this but better